+++ src/views/DaoInfo/components/StatsRibbon/StatsRibbon.tsx
@@ -13,7 +13,7 @@ import { AvailableFarms, INK } from "farms/AvailableFarms";
 import Box from "@mui/material/Box";
 
 // sets up a formatter so with toFormat on big numbers, we get thousands separators
-BigNumber.config({ FORMAT: { groupSeparator: ",", groupSize: 3 } });
+BigNumber.config({ FORMAT: { groupSeparator: ",", groupSize: 3, decimalSeparator: "." } });
 
 interface StatsRibbonProps {
   blockHeight?: string;
@@ -88,7 +88,7 @@ const StatsRibbon: React.FC<StatsRibbonProps> = ({ blockHeight }) => {
     fetchCurrentMarketcap();
   }, [BlockNum, fetchCurrentMarketcap]);
 
-  if (currentTvl === undefined) {
+  if (currentTvl === undefined || tokenPrices === undefined) {
     return <></>;
   }
 
@@ -100,7 +100,9 @@ const StatsRibbon: React.FC<StatsRibbonProps> = ({ blockHeight }) => {
           <Chip label={"Circulating INK: " + circulatingSupply} color="primary" variant="outlined" />
         </Tooltip>
         <Chip label={"Ecosystem TVL: $" + ecosystemTvl} color="primary" variant="outlined" />
-        <Chip label={"MarketCap: $" + currentMarketcap} color="primary" variant="outlined" />
+        <Tooltip title={"Fully Diluted: $" + new BigNumber(tokenPrices[INK]).times(88 * 1000000).toFormat(0)}>
+          <Chip label={"MarketCap: $" + currentMarketcap} color="primary" variant="outlined" />
+        </Tooltip>
         <Chip label={"24hr TXs: " + dailyTransactions} color="primary" variant="outlined" />
       </StyledBox>
     </>